+++ src/usecase/common.h
@@ -1,0 +1,16 @@
+#pragma once
+
+#include <string>
+#include <memory>
+#include "../model/Model.h"
+
+class Output {
+
+private:
+    std::string gameStatus;
+
+public:
+    void buildGameStatus(std::shared_ptr<Game> game);
+
+    std::string to_json();
+};
