+++ src/web/CreateGameCtrl.h
@@ -1,0 +1,16 @@
+#pragma once
+
+#include <drogon/HttpSimpleController.h>
+
+using namespace drogon;
+
+class CreateGameCtrl : public drogon::HttpSimpleController<CreateGameCtrl> {
+public:
+    void
+    asyncHandleHttpRequest(const HttpRequestPtr &req, std::function<void(const HttpResponsePtr &)> &&callback) override;
+
+    PATH_LIST_BEGIN
+        PATH_ADD("/guess_number_game:start", Post);
+    PATH_LIST_END
+};
+
